backend/server.py
from fastapi import FastAPI, HTTPException, Depends, status
from fastapi.security import HTTPBearer
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import os
from decouple import config
import uvicorn

app = FastAPI(title="CodeQuest API", version="1.0.0")

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure this properly in production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Environment variables
MONGO_URL = config("MONGO_URL", default="mongodb://localhost:27017/codequest")
GEMINI_API_KEY = config("GEMINI_API_KEY")
FIREBASE_PROJECT_ID = config("FIREBASE_PROJECT_ID")
JWT_SECRET_KEY = config("JWT_SECRET_KEY")

# Security
security = HTTPBearer()

# Pydantic models
class UserCreate(BaseModel):
    username: str
    email: str
    password: str

class UserLogin(BaseModel):
    email: str
    password: str

class Quest(BaseModel):
    id: str
    title: str
    description: str
    difficulty: str
    category: str
    xp_reward: int
    code_template: str
    expected_output: str
    test_cases: List[dict]

class UserProgress(BaseModel):
    user_id: str
    level: int
    xp: int
    completed_quests: List[str]
    current_quest: Optional[str] = None

# Basic routes
@app.get("/")
async def root():
    return {"message": "Welcome to CodeQuest API!"}

@app.get("/api/health")
async def health_check():
    return {"status": "healthy", "service": "CodeQuest Backend"}

# Authentication routes
@app.post("/api/auth/register")
async def register(user: UserCreate):
    # TODO: Implement user registration
    return {"message": "User registration endpoint - to be implemented"}

@app.post("/api/auth/login")
async def login(user: UserLogin):
    # TODO: Implement user login
    return {"message": "User login endpoint - to be implemented"}

@app.post("/api/auth/guest")
async def guest_login():
    # TODO: Implement guest session
    return {"message": "Guest login endpoint - to be implemented"}

# Quest routes
@app.get("/api/quests")
async def get_quests():
    # TODO: Implement quest retrieval
    return {"message": "Quest retrieval endpoint - to be implemented"}

@app.get("/api/quests/{quest_id}")
async def get_quest(quest_id: str):
    # TODO: Implement single quest retrieval
    return {"message": f"Quest {quest_id} retrieval endpoint - to be implemented"}

# Code execution routes
@app.post("/api/code/execute")
async def execute_code(code: dict):
    # TODO: Implement secure code execution
    return {"message": "Code execution endpoint - to be implemented"}

@app.post("/api/code/hint")
async def get_code_hint(request: dict):
    # TODO: Implement Gemini-powered hints
    return {"message": "Code hint endpoint - to be implemented"}

# User progress routes
@app.get("/api/user/progress")
async def get_user_progress():
    # TODO: Implement user progress retrieval
    return {"message": "User progress endpoint - to be implemented"}

@app.post("/api/user/progress")
async def update_user_progress(progress: UserProgress):
    # TODO: Implement user progress update
    return {"message": "User progress update endpoint - to be implemented"}

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8001)