backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
emergentintegrations==1.0.0
pymongo==4.6.1
uuid==1.30
python-decouple==3.8